Reorganize folder structure for admin pages
