+++ src/Geopilot.Frontend/src/pages/admin/organisations/organisations.tsx
@@ -1,13 +1,13 @@
 import { useTranslation } from "react-i18next";
 import { useCallback, useEffect, useState } from "react";
-import { Mandate, Organisation, User } from "../../api/apiInterfaces";
-import { useGeopilotAuth } from "../../auth";
+import { Mandate, Organisation, User } from "../../../api/apiInterfaces";
+import { useGeopilotAuth } from "../../../auth";
 import { Tooltip } from "@mui/material";
 import { GridActionsCellItem, GridColDef, GridRowId } from "@mui/x-data-grid";
-import { useControlledNavigate } from "../../components/controlledNavigate";
+import { useControlledNavigate } from "../../../components/controlledNavigate";
 import EditOutlinedIcon from "@mui/icons-material/EditOutlined";
-import GeopilotDataGrid from "../../components/geopilotDataGrid.tsx";
-import useFetch from "../../hooks/useFetch.ts";
+import GeopilotDataGrid from "../../../components/geopilotDataGrid.tsx";
+import useFetch from "../../../hooks/useFetch.ts";
 
 export const Organisations = () => {
   const { t } = useTranslation();
