+++ src/Geopilot.Frontend/src/pages/admin/users/users.tsx
@@ -1,13 +1,13 @@
 import { useTranslation } from "react-i18next";
 import { Tooltip } from "@mui/material";
-import { Organisation, User } from "../../api/apiInterfaces";
+import { Organisation, User } from "../../../api/apiInterfaces.ts";
 import { useCallback, useEffect, useState } from "react";
-import { useGeopilotAuth } from "../../auth";
+import { useGeopilotAuth } from "../../../auth/index.ts";
 import { GridActionsCellItem, GridColDef, GridRowId } from "@mui/x-data-grid";
 import EditOutlinedIcon from "@mui/icons-material/EditOutlined";
-import { useControlledNavigate } from "../../components/controlledNavigate";
-import GeopilotDataGrid from "../../components/geopilotDataGrid.tsx";
-import useFetch from "../../hooks/useFetch.ts";
+import { useControlledNavigate } from "../../../components/controlledNavigate/index.ts";
+import GeopilotDataGrid from "../../../components/geopilotDataGrid.tsx";
+import useFetch from "../../../hooks/useFetch.ts";
 
 export const Users = () => {
   const { t } = useTranslation();
